requirements.txt
# Core ML
xgboost==2.0.3
catboost==1.2.2
scikit-learn==1.4.0
numpy==1.26.3
pandas>=2.0.0

# Data sources
yfinance>=0.2.36,<2.0.0
alpaca-py==0.21.0

# API & Serving
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
jinja2>=3.1.0
aiofiles>=23.0.0
pandas

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
mlflow==2.10.2
optuna==3.5.0
schedule==1.2.1

# Monitoring & Validation
prometheus-client==0.19.0
psutil==5.9.8

# Testing
pytest<8, >=7.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Data processing
ta==0.11.0  # Technical Analysis library
pytz>=2022.5,<2024